refactor: auth module
