+++ libs/backend/src/infrastructure/auth/gql-auth.guard.ts
@@ -13,7 +13,7 @@ export class GqlAuthGuard extends AuthGuard('jwt') {
     return super.canActivate(new ExecutionContextHost([req]))
   }
 
-  handleRequest(err: any, user: any, info: any) {
+  handleRequest<User>(err: any, user: User, info: any, context: any) {
     // You can throw an exception based on either "info" or "err" arguments
     if (err || !user) {
       // throw err || new UnauthorizedException()
