+++ libs/modules/app/src/core/application/useCases/createApp/CreateAppRequest.ts
@@ -1,5 +1,7 @@
+import { User } from '@codelab/modules/user'
+
 export class CreateAppRequest {
   declare title: string
 
-  declare user?: User
+  declare user: User
 }
