+++ libs/modules/app/src/infrastructure/persistence/TypeOrmAppRepositoryAdapter.ts
@@ -2,11 +2,9 @@ import { plainToClass } from 'class-transformer'
 import { option as O } from 'fp-ts'
 import { Option } from 'fp-ts/Option'
 import { EntityRepository, Repository } from 'typeorm'
-import { TypeOrmApp } from '../../../../../backend/src/infrastructure/persistence/typeorm/entity/TypeOrmApp'
 import { Page } from '../../../../page/src/core/domain/page'
-import { FindAppBy } from '../../common/CommonTypes'
+import { AppsWhere, FindAppBy } from '../../common/CommonTypes'
 import { isId } from '../../common/utils'
-import { AppsWhere } from '../../common/CommonTypes'
 import { AppRepositoryPort } from '../../core/adapters/AppRepositoryPort'
 import { App } from '../../core/domain/app'
 import { TypeOrmApp, UUID } from '@codelab/backend'
