+++ libs/modules/app/src/presentation/controllers/AppCommandQueryAdapter.ts
@@ -2,15 +2,20 @@ import { Injectable, UseGuards } from '@nestjs/common'
 import { CommandBus, QueryBus } from '@nestjs/cqrs'
 import { Args, Mutation, Query, Resolver } from '@nestjs/graphql'
 import { classToPlain } from 'class-transformer'
-import { CurrentUser } from '../../../../../backend/src/infrastructure/auth/CurrentUser'
-import { GqlAuthGuard } from '../../../../../backend/src/infrastructure/auth/gql-auth.guard'
-import { TypeOrmApp } from '../../../../../backend/src/infrastructure/persistence/typeorm/entity/TypeOrmApp'
 import { CreateAppCommand } from '../../core/application/commands/CreateAppCommand'
+import { DeleteAppCommand } from '../../core/application/commands/DeleteAppCommand'
 import { GetAppsQuery } from '../../core/application/commands/GetAppsQuery'
 import { AppDto } from '../../core/application/useCases/AppDto'
 import { CreateAppInput } from '../../core/application/useCases/createApp/CreateAppInput'
+import { DeleteAppRequest } from '../../core/application/useCases/deleteApp/DeleteAppRequest'
 import { GetAppsRequest } from '../../core/application/useCases/getApps/GetAppsRequest'
-import { CommandQueryBusPort, UseCaseRequestPort } from '@codelab/backend'
+import {
+  CommandQueryBusPort,
+  CurrentUser,
+  GqlAuthGuard,
+  TypeOrmApp,
+  UseCaseRequestPort,
+} from '@codelab/backend'
 import { User } from '@codelab/modules/user'
 
 @Resolver(() => TypeOrmApp)
@@ -25,10 +30,10 @@ export class AppCommandQueryAdapter implements CommandQueryBusPort {
   @UseGuards(GqlAuthGuard)
   async createApp(
     @Args('input') input: CreateAppInput,
-    @CurrentUser() userId: string,
+    @CurrentUser() user: User,
   ) {
     const results = await this.commandBus.execute(
-      new CreateAppCommand({ ...input, userId }),
+      new CreateAppCommand({ ...input, user }),
     )
 
     return results.toPlain()
@@ -44,7 +49,7 @@ export class AppCommandQueryAdapter implements CommandQueryBusPort {
     return classToPlain(results)
   }
 
-  @Mutation((returns) => AppUseCaseDto)
+  @Mutation((returns) => AppDto)
   @UseGuards(GqlAuthGuard)
   async deleteApp(@Args('request') request: DeleteAppRequest) {
     const result = await this.commandBus.execute(new DeleteAppCommand(request))
