+++ libs/modules/user/src/core/application/useCases/validateUser/ValidateUserService.ts
